Revert some settings
