--- .worker_env
@@ -1,3 +1,0 @@
-CELERY_BROKER_URL=redis://redis:6379/0
-CELERY_RESULT_BACKEND=redis://redis:6379/0
-C_FORCE_ROOT=true
