--- .worker_env.example
@@ -1,5 +1,0 @@
-CELERY_BROKER_URL=redis://redis:6379/0
-CELERY_RESULT_BACKEND=redis://redis:6379/0
-C_FORCE_ROOT=true
-DATASET_DIR=/datasets
-FILE_DATA=/file_data
